sales_agents_guardrails/requirements.txt
# Sales Agents with Guardrails - Dependencies
# Core AI and Agent Framework
agents>=0.1.0
openai>=1.0.0

# Data Validation and Models
pydantic>=2.0.0

# Environment Configuration
python-dotenv>=1.0.0

# Email Services
sendgrid>=6.10.0

# HTTP Client for API calls
httpx>=0.25.0

# Async Support
asyncio-mqtt>=0.13.0

# Development and Testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Documentation (optional)
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
